[network] - merge networking.go into network.go
